+++ database/migrations/2025_08_19_210507_add_url_to_actions_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! Schema::hasTable('actions')) {
+            return;
+        }
+
         Schema::table('actions', function (Blueprint $table) {
             $table->string('url')->nullable()->after('note');
             //
@@ -22,6 +26,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! Schema::hasTable('actions')) {
+            return;
+        }
+
         Schema::table('actions', function (Blueprint $table) {
             $table->dropColumn('url');
         });
